Move TemplateWithUrl -> CustomTemplate transformation to useCase
