+++ i18n/es.po
@@ -1,7 +1,7 @@
 msgid ""
 msgstr ""
 "Project-Id-Version: Bulk Load\n"
-"POT-Creation-Date: 2022-04-04T12:06:50.406Z\n"
+"POT-Creation-Date: 2022-04-06T09:18:23.150Z\n"
 "Language: es\n"
 "MIME-Version: 1.0\n"
 "Content-Type: text/plain; charset=UTF-8\n"
@@ -386,6 +386,14 @@ msgstr "Tipo"
 
 msgid "Custom"
 msgstr ""
+
+#, fuzzy
+msgid "Created"
+msgstr "Crear"
+
+#, fuzzy
+msgid "Last Updated"
+msgstr "Actualizados"
 
 #, fuzzy
 msgid "Sharing Settings"
