+++ i18n/fr.po
@@ -1,7 +1,7 @@
 msgid ""
 msgstr ""
 "Project-Id-Version: Bulk Load App\n"
-"POT-Creation-Date: 2022-04-04T12:06:50.406Z\n"
+"POT-Creation-Date: 2022-04-06T09:18:23.150Z\n"
 "Language: fr\n"
 "MIME-Version: 1.0\n"
 "Content-Type: text/plain; charset=UTF-8\n"
@@ -364,6 +364,14 @@ msgstr ""
 
 msgid "Custom"
 msgstr ""
+
+#, fuzzy
+msgid "Created"
+msgstr "Créer"
+
+#, fuzzy
+msgid "Last Updated"
+msgstr "Actualisé"
 
 #, fuzzy
 msgid "Sharing Settings"
