+++ src/webapp/components/settings/SettingsFields.tsx
@@ -38,7 +38,7 @@ export default function SettingsFields(props: SettingsFieldsProps & CustomTempla
     const [isDataSetDataElementsFilterDialogVisible, showDataSetDataElementsFilterDialog] = useState<boolean>(false);
     const [isProgramStageFilterDialogVisible, showProgramStageFilterDialog] = useState<boolean>(false);
     const [isModuleTemplateDialogVisible, showModuleTemplateDialog] = useState<boolean>(false);
-    const [isCustomTemplatesDialogVisible, showCustomTemplatesDialog] = useState<boolean>(true);
+    const [isCustomTemplatesDialogVisible, showCustomTemplatesDialog] = useState<boolean>(false);
 
     const setModel = useCallback(
         (model: Model) => {
